Add namespaces and update PHP support to 5.3
