+++ composer.json
@@ -6,13 +6,15 @@
     "homepage": "https://github.com/pusher/pusher-php-server",
     "license": "MIT",
     "require": {
-        "php": ">=5.2",
+        "php": ">=5.3.3",
         "ext-curl": "*"
     },
     "require-dev": {
-        "phpunit/phpunit": "3.7.*"
+        "phpunit/phpunit": "~4.0"
     },
     "autoload": {
-        "classmap": [ "lib/" ]
+        "psr-4": {
+            "Pusher\\": "src/"
+        }
     }
 }
