+++ src/Exceptions/PusherException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Pusher\Exceptions;
+
+class PusherException extends Exception {}
