+++ src/Pusher.php
@@ -1,6 +1,6 @@
 <?php
 
-/* 
+/*
 		Pusher PHP Library
 	/////////////////////////////////
 	PHP library for the Pusher API.
@@ -20,33 +20,9 @@
 		+ Phil Leggetter (phil@leggetter.co.uk)
 */
 
-class PusherException extends Exception
-{
-}
+namespace Pusher;
 
-class PusherInstance {
-	
-	private static $instance = null;
-	private static $app_id	= '';
-	private static $secret	= '';
-	private static $api_key = '';
-	
-	private function __construct() { }
-	private function __clone() { }
-	
-	public static function get_pusher()
-	{
-		if (self::$instance !== null) return self::$instance;
-
-		self::$instance = new Pusher(
-			self::$api_key, 
-			self::$secret, 
-			self::$app_id
-		);
-
-		return self::$instance;
-	}
-}
+use Pusher\Exceptions\PusherException;
 
 class Pusher
 {
@@ -62,11 +38,11 @@ class Pusher
 	private $logger = null;
 
 	/**
-	 * PHP5 Constructor. 
-	 * 
-	 * Initializes a new Pusher instance with key, secret , app ID and channel. 
+	 * PHP5 Constructor.
+	 *
+	 * Initializes a new Pusher instance with key, secret , app ID and channel.
 	 * You can optionally turn on debugging for all requests by setting debug to true.
-	 * 
+	 *
 	 * @param string $auth_key
 	 * @param string $secret
 	 * @param int $app_id
@@ -96,15 +72,15 @@ class Pusher
 		if( !is_null( $host ) ) {
 			$match = null;
 			preg_match("/(http[s]?)\:\/\/(.*)/", $host, $match);
-			
+
 			if( count( $match ) === 3 ) {
 				$this->settings[ 'scheme' ] = $match[ 1 ];
 				$host = $match[ 2 ];
 			}
-			
+
 			$this->settings[ 'host' ] = $host;
-			
-			$this->log( 'Legacy $host parameter provided: ' . 
+
+			$this->log( 'Legacy $host parameter provided: ' .
 									$this->settings[ 'scheme' ] + ' host: ' + $this->settings[ 'host' ] );
 		}
 
@@ -115,14 +91,14 @@ class Pusher
 		if( !is_null( $timeout ) ) {
 			$options[ 'timeout' ] = $timeout;
 		}
-		
+
 		/** End backward compatibility with old constructor **/
-		
+
 		if( isset( $options[ 'encrypted' ] ) &&
 				$options[ 'encrypted' ] === true &&
 				!isset( $options[ 'scheme' ] ) &&
 				!isset( $options[ 'port' ] ) ) {
-						
+
 			$options[ 'scheme' ] = 'https';
 			$options[ 'port' ] = 443;
 		}
@@ -138,9 +114,9 @@ class Pusher
 				$this->settings[ $key ] = $value;
 			}
 		}
-		
+
 		// ensure host doesn't have a scheme prefix
-		$this->settings[ 'host' ] = 
+		$this->settings[ 'host' ] =
 			preg_replace( '/http[s]?\:\/\//', '', $this->settings[ 'host' ], 1 );
 	}
 
@@ -184,7 +160,7 @@ class Pusher
 		}
 
 	}
-	
+
 	/**
 	 * Utility function used to create the curl object with common settings
 	 */
@@ -199,23 +175,23 @@ class Pusher
 			$query_params);
 
 		$full_url = $this->settings['scheme'] . '://' .
-								$this->settings['host'] . ':' . 
+								$this->settings['host'] . ':' .
 								$this->settings['port'] . $s_url . '?' . $signed_query;
 
 		$this->log( 'curl_init( ' . $full_url . ' )' );
-		
+
 		# Set cURL opts and execute request
 		$ch = curl_init();
 		if ( $ch === false )
 		{
 			throw new PusherException('Could not initialise cURL!');
 		}
-		
+
 		curl_setopt( $ch, CURLOPT_URL, $full_url );
 		curl_setopt( $ch, CURLOPT_HTTPHEADER, array ( "Content-Type: application/json" ) );
 		curl_setopt( $ch, CURLOPT_RETURNTRANSFER, 1 );
 		curl_setopt( $ch, CURLOPT_TIMEOUT, $this->settings['timeout'] );
-		
+
 		return $ch;
 	}
 
@@ -229,7 +205,7 @@ class Pusher
 		$response[ 'status' ] = curl_getinfo($ch, CURLINFO_HTTP_CODE);
 
 		$this->log( 'exec_curl response: ' . print_r( $response, true ) );
-		
+
 		if( $response[ 'body' ] === false ) {
 			$this->log( 'exec_curl error: ' . curl_error( $ch ) );
 		}
@@ -238,7 +214,7 @@ class Pusher
 
 		return $response;
 	}
-	
+
 	/**
 	 * Build the required HMAC'd auth string
 	 *
@@ -253,27 +229,27 @@ class Pusher
 	 */
 	public static function build_auth_query_string($auth_key, $auth_secret, $request_method, $request_path,
 		$query_params = array(), $auth_version = '1.0', $auth_timestamp = null)
-	{ 
+	{
 		$params = array();
 		$params['auth_key'] = $auth_key;
 		$params['auth_timestamp'] = (is_null($auth_timestamp)?time() : $auth_timestamp);
 		$params['auth_version'] = $auth_version;
-		
+
 		$params = array_merge($params, $query_params);
 		ksort($params);
-		
+
 		$string_to_sign = "$request_method\n" . $request_path . "\n" . Pusher::array_implode( '=', '&', $params );
 
 		$auth_signature = hash_hmac( 'sha256', $string_to_sign, $auth_secret, false );
-		
+
 		$params['auth_signature'] = $auth_signature;
 		ksort($params);
-		
+
 		$auth_query_string = Pusher::array_implode( '=', '&', $params );
-		
+
 		return $auth_query_string;
 	}
-	
+
 	/**
 	 * Implode an array with the key and value pair giving
 	 * a glue, a separator between pairs and the array
@@ -291,12 +267,12 @@ class Pusher
 							$val = implode( ',', $val );
 					$string[] = "{$key}{$glue}{$val}";
 
-			}		 
+			}
 			return implode( $separator, $string );
 	}
 
 	/**
-	 * Trigger an event by providing event name and payload. 
+	 * Trigger an event by providing event name and payload.
 	 * Optionally provide a socket ID to exclude a client (most likely the sender).
 	 *
 	 * @param array $channels An array of channel names to publish the event on.
@@ -318,9 +294,9 @@ class Pusher
 		}
 
 		$query_params = array();
-		
-		$s_url = $this->settings['base_path'] . '/events';		
-		
+
+		$s_url = $this->settings['base_path'] . '/events';
+
 		$data_encoded = $already_encoded ? $data : json_encode( $data );
 
 		$post_params = array();
@@ -360,7 +336,7 @@ class Pusher
 		}
 
 	}
-	
+
 	/**
 	 *	Fetch channel information for a specific channel.
 	 *
@@ -371,7 +347,7 @@ class Pusher
 	public function get_channel_info($channel, $params = array() )
 	{
 		$response = $this->get( '/channels/' . $channel, $params );
-		
+
 		if( $response[ 'status' ] == 200)
 		{
 			$response = json_decode( $response[ 'body' ] );
@@ -380,13 +356,13 @@ class Pusher
 		{
 			$response = false;
 		}
-		
+
 		return $response;
 	}
-	
+
 	/**
 	 * Fetch a list containing all channels
-	 * 
+	 *
 	 * @param array $params Additional parameters for the query e.g. $params = array( 'info' => 'connection_count' )
 	 *
 	 * @return array
@@ -394,7 +370,7 @@ class Pusher
 	public function get_channels($params = array())
 	{
 		$response = $this->get( '/channels', $params );
-		
+
 		if( $response[ 'status' ] == 200)
 		{
 			$response = json_decode( $response[ 'body' ] );
@@ -404,26 +380,26 @@ class Pusher
 		{
 			$response = false;
 		}
-		
+
 		return $response;
 	}
 
 	/**
 	 * GET arbitrary REST API resource using a synchronous http client.
 	 * All request signing is handled automatically.
-	 *	
+	 *
 	 * @param string path Path excluding /apps/APP_ID
 	 * @param params array API params (see http://pusher.com/docs/rest_api)
 	 *
 	 * @return See Pusher API docs
 	 */
 	public function get( $path, $params = array() ) {
-		$s_url = $this->settings['base_path'] . $path;	
+		$s_url = $this->settings['base_path'] . $path;
 
 		$ch = $this->create_curl( $s_url, 'GET', $params );
 
 		$response = $this->exec_curl( $ch );
-		
+
 		if( $response[ 'status' ] == 200)
 		{
 			$response[ 'result' ] = json_decode( $response[ 'body' ], true );
@@ -432,13 +408,13 @@ class Pusher
 		{
 			$response = false;
 		}
-		
+
 		return $response;
 	}
 
 	/**
 	 * Creates a socket signature
-	 * 
+	 *
 	 * @param int $socket_id
 	 * @param string $custom_data
 	 * @return string
