+++ test/acceptance/channelQueryTest.php
@@ -1,5 +1,7 @@
 <?php
-	
+
+use Pusher\Pusher;
+
 	class PusherChannelQueryTest extends PHPUnit_Framework_TestCase
 	{
 
@@ -12,29 +14,29 @@
 		public function testChannelInfo()
 		{
 			$response = $this->pusher->get_channel_info('channel-test');
-			
+
 			//print_r( $response );
-			
+
 			$this->assertObjectHasAttribute('occupied', $response, 'class has occupied attribute');
 		}
-		
+
 		public function testChannelList()
 		{
 			$result = $this->pusher->get_channels();
 			$channels = $result->channels;
-			
+
 			 // print_r( $channels );
-			
+
 			foreach( $channels as $channel_name => $channel_info ) {
         echo( "channel_name: $channel_name\n");
         echo( 'channel_info: ' );
         print_r( $channel_info );
         echo( "\n\n");
       }
-			
+
 			$this->assertTrue( is_array($channels), 'channels is an array' );
 		}
-		
+
 		public function testFilterByPrefixNoChannels()
 		{
 			$options = array(
@@ -45,9 +47,9 @@
 // print_r( $result );
 
 		  $channels = $result->channels;
-		  
+
 		  // print_r( $channels );
-		  
+
 			$this->assertTrue( is_array($channels), 'channels is an array' );
 			$this->assertEquals( 0, count( $channels ), 'should be an empty array' );
 		}
@@ -62,9 +64,9 @@
 // print_r( $result );
 
 		  $channels = $result->channels;
-		  
+
 		  // print_r( $channels );
-		  
+
 			$this->assertEquals( 1, count( $channels ), 'channels have a single test-channel present. For this test to pass you must have your API Access setting open for the application you are testing against' );
 		}
 
@@ -76,7 +78,7 @@
 				'info' => 'user_count'
 			);
 		  $result = $this->pusher->get_channels( $options );
-		  
+
 			$this->assertFalse( $result, 'query should fail' );
 		}
 
@@ -119,10 +121,10 @@
 			$this->assertEquals( $response[ 'status' ], 200 );
 
 			$result = $response[ 'result' ];
-			
+
 			$this->assertArrayHasKey('occupied', $result, 'class has occupied attribute');
 		}
-		
+
 	}
 
 ?>
