+++ test/test_includes.php
@@ -14,6 +14,6 @@ else
   define('PUSHERAPP_HOST', 'http://api.pusherapp.com');
 }
 
-require_once($dir . '/../lib/Pusher.php');
+require_once($dir . '/../src/Pusher.php');
 
 require_once( $dir . '/TestLogger.php' );
