+++ test/unit/authQueryStringTest.php
@@ -1,33 +1,35 @@
 <?php
-	
+
+use Pusher\Pusher;
+
 	class PusherAuthQueryString extends PHPUnit_Framework_TestCase
 	{
 
 		protected function setUp()
 		{
-			$this->pusher = new Pusher('thisisaauthkey', 'thisisasecret', 1, true); 
+			$this->pusher = new Pusher('thisisaauthkey', 'thisisasecret', 1, true);
 		}
-		
+
 		public function testArrayImplode()
 		{
 		  $val = array('testKey' => 'testValue');
-		  
+
 		  $expected = 'testKey=testValue';
 		  $actual = Pusher::array_implode( '=', '&', $val );
-		  
-		  $this->assertEquals($actual, 
+
+		  $this->assertEquals($actual,
 				$expected,
 				'auth signature valid');
 		}
-		
+
 		public function testArrayImplodeWithTwoValues()
 		{
 		  $val = array('testKey' => 'testValue', 'testKey2' => 'testValue2');
-		  
+
 		  $expected = 'testKey=testValue&testKey2=testValue2';
 		  $actual = Pusher::array_implode( '=', '&', $val );
-		  
-		  $this->assertEquals($actual, 
+
+		  $this->assertEquals($actual,
 				$expected,
 				'auth signature valid');
 		}
@@ -51,17 +53,17 @@
   		  $query_params,
   		  $auth_version,
   		  $time);
-  		  
+
   		$expected_to_sign = "POST\n$request_path\nauth_key=$auth_key&auth_timestamp=$time&auth_version=$auth_version&name=an_event";
   		$expected_auth_signature = hash_hmac( 'sha256', $expected_to_sign, $auth_secret, false );
   		$expected_query_string = "auth_key=$auth_key&auth_signature=$expected_auth_signature&auth_timestamp=$time&auth_version=$auth_version&name=an_event";
-			
-			
-			$this->assertEquals($auth_query_string, 
+
+
+			$this->assertEquals($auth_query_string,
 				$expected_query_string,
 				'auth signature valid');
 		}
-		
+
 	}
 
 ?>
